requirements.txt
Flask==2.3.3
Jinja2==3.1.2
Pillow==10.0.1
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
psutil==5.9.6 